% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateAnalysis.R
\name{bivariateAnalysis}
\alias{bivariateAnalysis}
\title{A bivariateAnalysis Function}
\usage{
bivariateAnalysis(dat, na.rm = T, plot = F, warning = F)
}
\arguments{
\item{dat}{data in data.frame format}

\item{na.rm}{TRUE if missing data is NOT to be used in analysis}

\item{plot}{TRUE if you wish to produce bivariate plots}

\item{warning}{TRUE to show all warnings generated from statistical analysis}
}
\value{
Return statistics from bivariate analysis.
}
\description{
This function performs appropriate bivariate analysis based on the data type of dependent (DV) and independent (IV) variables. Pearson
correlation is used for numerical data, and Kendall Tau if data is ordinal. Chi-square test of association is used for nominal data, and when nominal data
are less than 5 in any category, the Fisher test is used instead. To analyse association between numerical and nominal data,
we first determine if the data distribution is normal using Shapiro-Wilk test. For normal distributed data, we use t-test for
two levels nominal data, and one-way ANOVA for 3-levels or more nominal data. We use non-parametric tests i.e. Mann-Whitney
for nominal data with 2 levels, and Kruskal-Wallis when there are more than 2 levels nominal data.
}
\examples{
data(iris)
result = bivariateAnalysis(iris)
}
\references{
\url{http://www.biostathandbook.com/testchoice.html}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{bivariate}
