% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkImpute.R
\name{checkImpute}
\alias{checkImpute}
\title{A checkImpute Function}
\usage{
checkImpute(before, after, sig = 0.05, plot = F)
}
\arguments{
\item{before}{data frame before perform imputation.}

\item{after}{data frame after perform imputation.}

\item{sig}{significant value threshold.}

\item{plot}{TRUE to plot before and after imputation data with significant different distribution.}
}
\value{
Return a list of variables that have different distribution statistically after imputation.
}
\description{
This function performs statistical tests to check if two data distributions generated from before and after imputation
are the same.
}
\examples{
library(FactoMineR)
data(wine)
set.seed(100)
for(i in 2:ncol(wine))
{
#Generate missing data
missing = sample.int(nrow(wine), round(0.2*nrow(wine)))
wine[missing, i] = NA
}

#Impute using mean
impute.wine = wine
for(i in 2:ncol(wine))
{
  impute.wine[which(is.na(wine[, i])), i] = mean(wine[, i], na.rm = TRUE)
}

checkImpute(wine, impute.wine, sig = 0.05)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{imputation}
