% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlot.R
\name{clusterPlot}
\alias{clusterPlot}
\title{A clusterPlot Function}
\usage{
clusterPlot(x, y, title = "Cluster plot", use.PCA = FALSE, color.blind = T)
}
\arguments{
\item{x}{data in a data.frame format}

\item{y}{cluster results}

\item{title}{title for the cluster plot}

\item{use.PCA}{TRUE to use PCA to tranform data for plotting}

\item{color.blind}{TRUE if colors for the colour blind palette to be used}
}
\description{
This function produces a cluster plot using the first two columns as axes. Otherwise sepcify use.PCA
to TRUE to use PCA to tranform data for plotting.
}
\examples{
data(iris)
cm = kmeans(iris[, -5], centers = 3)
clusterPlot(iris[, -5], y = factor(cm$cluster))
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{cluster}
