% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMR.R
\name{MRMR}
\alias{MRMR}
\title{A MRMR Function}
\usage{
MRMR(x, y, no.bin = 5, criteria = "MID", no.of.variable = 5)
}
\arguments{
\item{x}{independent variable in data.frame.}

\item{y}{dependent variable.}

\item{no.bin}{number of bin used if variables are continuous.}

\item{criteria}{feature selection scheme to use. 'MID' for Mutual Information Difference
criterion, or 'MIQ' for Mutual Information Quotient criterion.}

\item{no.of.variable}{Number of variable to be selected.}
}
\value{
Return the selected features.
}
\description{
This function performs feature selection based on Mutual Information using the maximum relevance
minimum redundancy (MRMR).
}
\examples{
data(iris)
selected.feature = MRMR(iris[, -5], iris[, 5], no.bin = 3, criteria = "MID")
}
\references{
\url{https://link.springer.com/article/10.1007/s10489-018-1305-0}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{feature}
\keyword{selection}
