% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{outlierDetection}
\alias{outlierDetection}
\title{A outlierDetection Function}
\usage{
outlierDetection(
  dat,
  grubb = FALSE,
  basic = FALSE,
  beforePlot = F,
  outlierPlot = F,
  ...
)
}
\arguments{
\item{dat}{a data frame for data values}

\item{grubb}{Grubb method performs grubbs' test to detect outliers.}

\item{basic}{Basic method detects data that lies outside the specified range, given by alpha}

\item{beforePlot}{plot data distribution of the data}

\item{outlierPlot}{pot data distribution when outliers are removed}

\item{...}{other arguments}
}
\value{
Return a list of outliers containing the data id and their values.
}
\description{
This function return a list of outliers detected.
}
\examples{
data(iris)
outlierDetection(iris, grubb = TRUE)
}
\references{
\url{https://towardsdatascience.com/a-brief-overview-of-outlier-detection-techniques-1e0b2c19e561}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{outlier}
