% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{A plotMissing Function}
\usage{
plotMissing(dat, plot = T)
}
\arguments{
\item{dat}{data in data frame format.}

\item{plot}{TRUE to plot percentage of missing data}
}
\value{
Return percentage of data missing in each variable
}
\description{
This function produces a plot showing percentage of data missing in each variable.
}
\examples{
library( FactoMineR)
data(wine)
set.seed(100)
for(i in 2:ncol(wine))
{
#Generate missing data
missing = sample.int(nrow(wine), round(0.2*nrow(wine)))
wine[missing, i] = NA
}
plotMissing(wine)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{data}
\keyword{missing}
