% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateAnalysis.R
\name{univariateAnalysis}
\alias{univariateAnalysis}
\title{A univariateAnalysis Function}
\usage{
univariateAnalysis(
  dat,
  hist = FALSE,
  boxplot = FALSE,
  qqnorm = FALSE,
  shapiro = FALSE
)
}
\arguments{
\item{dat}{data in data.frame format}

\item{hist}{a plot of histogram}

\item{boxplot}{a box plot showing distribution of the variable}

\item{qqnorm}{a qnantile quantile plot}

\item{shapiro}{a  Shapiro-Wilk normality test.}
}
\value{
Return plots from univariate analysis.
}
\description{
This function produces univariate plots including histogram with density plot, box plot, q-q plot, and calculate
the Shapiro-Wilk statistic for numeric data.
}
\examples{
data(iris)
univariateAnalysis (iris, qqnorm  = TRUE)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{univariate}
